# 64-bit compile fix (see http://www.wildfiregames.com/forum/index.php?showtopic=12448)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7031 3db68df2-c116-0410-a063-a993310a9797

+++ source/collada/XMLFix.cpp
@@ -174,7 +174,7 @@ void FixBrokenXML(const char* text, const char** out, size_t* outSize)
 		// Reserialising the document, then parsing it again inside FCollada, is a bit ugly;
 		// but it's the only way I can see to make it work through FCollada's public API
 		xmlChar* mem = NULL;
-		int size = -1;
+		intptr_t size = -1;
 		xmlDocDumpFormatMemory(doc, &mem, &size, 0);
 		*out = (const char*)mem;
 		*outSize = size;
